requirements.txt
slackbot
PyYAML
